backend/requirements.txt
Flask
flask_cors
python-dotenv
mysql-connector-python
langchain
langchain-groq
deep-translator

